refactor(email): make attachments available for general send email
