+++ packages/backend/src/apps/postman/common/parameters.ts
@@ -3,6 +3,8 @@ import { IField } from '@plumber/types'
 import validator from 'email-validator'
 import { z } from 'zod'
 
+import { parseS3Id } from '@/helpers/s3'
+
 function recipientStringToArray(value: string) {
   return value
     .split(',')
@@ -50,6 +52,14 @@ export const transactionalEmailFields: IField[] = [
     required: true,
     variables: true,
   },
+  {
+    label: 'Attachments',
+    key: 'attachments',
+    type: 'multiselect' as const,
+    required: false,
+    variables: true,
+    variableTypes: ['file'],
+  },
 ]
 
 export const transactionalEmailSchema = z.object({
@@ -76,4 +86,22 @@ export const transactionalEmailSchema = z.object({
     return value.trim() === '' ? undefined : value.trim()
   }, z.string().email().optional()),
   senderName: z.string().min(1).trim(),
+  attachments: z.array(z.string()).transform((array, context) => {
+    const result: string[] = []
+    for (const value of array) {
+      // Account for optional attachment fields with no response.
+      if (!value) {
+        continue
+      }
+      if (!parseS3Id(value)) {
+        context.addIssue({
+          code: z.ZodIssueCode.custom,
+          message: `${value} is not a S3 ID.`,
+        })
+        return z.NEVER
+      }
+      result.push(value)
+    }
+    return result
+  }),
 })
